// runtime/save-data/types.ts
export type AccountSeed = {
  uid: string;
  username: string;
  nickname: string;
};

export type Account = AccountSeed & {
  id: number;
};

export type SaveSlot = {
  index: number;
  title: string;
  datetime: string;
  data?: unknown;
  status: string;
};

export type SavePayload = {
  uid: string;
  gameid: string;
  index: number;
  title: string;
  data: string;
};
// ...
export type Wallet = {
  balance: number;
  totalPaid: number;
  totalRecharged: number;
};

export type RechargeWalletRequest = {
  uid: string;
  amount: number;
};

export type RechargeWalletResult = Wallet & {
  amount: number;
};

export type BuyPropRequest = {
  uid: string;
  gameId?: string;
  slotIndex?: number;
  propId: number;
  count: number;
  price: number;
  tag: number;
};

export type BuyPropResult = {
  propId: number;
  count: number;
  tag: number;
  balance: number;
};